Only delete files, not links

+++ app/workers/zipfulldata.rb
@@ -62,6 +62,7 @@ class Zipfulldata
         Rails.root.join("public/data/zip/#{dump_file_name}.zip"),
         link_path)
         
+      puts Rails.root.join("public/data/zip/#{dump_file_name}.zip")
       # everything went OK, now delete old zips
       delete_old_zips
 
@@ -257,9 +258,9 @@ TXT
   end
 
   def delete_old_zips
-    forbidden_files = [link_path, zip_fs_path, Rails.root.join("data/annotation.zip")]
+    forbidden_files = [link_path, zip_fs_path, Rails.root.join("data/annotation.zip"), Rails.root.join("public/data/zip/#{dump_file_name}.zip")] 
     Dir[Rails.root.join('public/data/zip/*.zip')].each do |f|
-      if not forbidden_files.include? f
+      if (not forbidden_files.include? f) and (File.ftype(f) == "file")
         File.delete(f)
       end
     end
